@@ -1,26 +1,22 @@
+import 'package:firebase_database/firebase_database.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
-mixin GetUserInfo{
-  gettingData() {
-    StreamBuilder<QuerySnapshot> build = StreamBuilder<QuerySnapshot>(
-      stream: Firestore.instance.collection('users').snapshots(),
-      builder: (BuildContext context, AsyncSnapshot<QuerySnapshot> snapshot) {
-        switch (snapshot.connectionState) {
-          case ConnectionState.waiting:
-            return new Center(child: new CircularProgressIndicator());
-          default:
-            return ListView(
-                children:
-                snapshot.data.documents.map((DocumentSnapshot document) {
-                  return ListTile(
-                    title: Text(document['email']),
 
-                  );
-                }).toList());
-        }
-      },
-    );
-    return build;
+class GetUserInfo {
+  String _id;
+  String _name;
+  String _email;
+
+  //GetUserInfo(this._id,this._name, this._email);
+
+  GetUserInfo.fromSnapshot(DataSnapshot snapshot) {
+    _id = snapshot.key;
+    _name = snapshot.value['name'];
+    _email = snapshot.value['email'];
   }
+  String get name => _name;
+
+  String get email => _email;
+
+  String get id => _id;
 }
